pyproject.toml
[project]
name = "boutique"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.24.0",
    "python-dotenv>=1.1.1",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]